A Go IRC Bot.
This is my latest pet project.

A lot of the code is based on thoj's "Go-IRC-Client-Library"/irc.go:
http://github.com/thoj/Go-IRC-Client-Library/blob/master/irc.go

but crammed into one file, just to give me a chance to learn Go and Go
sockets a bit better. The code is really crappy right now, but it does
connect and PONG.